backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jacob Fresh <jacob@fresh.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
pre-commit = "^2.11.1"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
ipython = "^7.21.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
